refactor: models 위치 변경
