+++ src/refactoring/entities/cart/cart.ts
@@ -1,4 +1,4 @@
-import { Product } from "./product";
+import { Product } from '../product';
 
 export interface CartItem {
   product: Product;
