+++ src/refactoring/entities/index.ts
@@ -1,3 +1,4 @@
 export type { Product, Discount } from './product';
 export type { Coupon } from './coupon';
-export type { CartItem } from './cart';
+export type { CartItem } from './cart/cart';
+export * from './cart/lib';
